Add front-end changes for registration, scss, and tags
