+++ backend/setup/env_var.sh
@@ -1,0 +1,3 @@
+export FLASK_APP=simpletodolist.py
+export FLASK_DEBUG=1
+export SECRET_CODE="simple-todo-secret-code"
