+++ frontend/src/assets/scss/_dashboard.scss
@@ -1,0 +1,173 @@
+$primary-font: 'Source Sans Pro', sans-serif;
+
+.login-label {
+  font-weight: 700;
+  margin: 20px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+} 
+
+.login-username, .login-password {
+  margin: 0 10px;
+}
+
+.login-register, .login-submit {
+  width: 100px;
+}
+
+.register-label {
+  width: 150px;
+  text-align: left;
+}
+
+body {
+  background-color: #daf7dc;
+}
+
+#app {
+  font-family: $primary-font;
+}
+
+input {
+  font-family: $primary-font;
+  padding: 10px;
+  margin-right: 20px;
+}
+
+ul {
+  list-style-type: none;
+	margin: 0 auto;
+}
+
+h2, h3 {
+  padding: 5px;
+}
+
+h2:focus, h3:focus {
+ background-color: #fff; 
+}
+
+.main-dashboard {
+  text-align: left;
+  max-width: 800px;
+  margin: 0 auto;
+
+  .icon {
+    font-size: 30px;
+    vertical-align: middle;
+    color: #42273b;
+    margin: 0 10px;
+    text-shadow: 5px 5px rgba(0,0,0,0.15);
+  }
+
+  .ion-md-close {
+    font-size: 20px !important;
+    margin: 0 5px !important;
+  }
+
+  .add-new-task-container {
+    text-align: left;
+
+    #new-task-description {
+      width: 500px;
+    }
+
+  }
+
+  .task-container {
+    background-color: #abc8c0;
+    color: #70566d;
+    margin: 60px 0;
+    padding: 20px;
+
+    .task-title {
+      margin: 0;
+      font-weight: 700;
+      font-size: 30px;
+    }
+
+    .task-description {
+      margin: 0;
+      font-weight: 300;
+      font-size: 25px;
+    }
+  }
+
+
+  .specific-task {
+    display: inline-block;
+  }
+  .task-buttons {
+    float: right;
+  }
+
+  .add-new-todo-container {
+    .new-todo-error-message {
+      font-weight: 700;
+      color: #daf7dc;
+    }
+  }
+
+  .todo-container {
+    display: inline-block;
+    width: 84%;
+    margin-top: 20px;
+
+    .todo-delete {
+      float: right;
+    }
+
+    .todo-description {
+      font-weight: 300;
+      font-size: 20px;
+      display: inline;
+    }
+
+    .todo-expdate {
+      font-weight: 700;
+      font-size: 15px;
+      display: inline;
+    }
+
+  }
+
+  .tag-names-container {
+    margin-top: 20px;
+    margin-left: 40px;
+  }
+
+  .add-new-todo {
+    display: block;
+    clear: both;
+  }
+
+  input[type=checkbox]:checked + h2.todo-description{
+    text-decoration: line-through;
+  }
+
+  .tag-name {
+    background-color: #daf7dc;
+    margin: 0 5px;
+    padding: 5px;
+  }
+
+	.new-todo-tags {
+		position:relative;
+	}
+
+	.show-top-tags {
+		display: none;
+		text-align: center;
+		width: 500px;
+		padding: 20px;
+	  background: #daf7dc;
+	  opacity:.9;
+	}
+
+	.new-todo-tags:focus ~ .show-top-tags {
+	  display: block;
+	  position: absolute;
+	}
+}
+
